pyproject.toml
[tool.poetry]
name = "tutorialproject"
version = "0.1.0"
description = ""
authors = ["smwasifzamanomee <wasifbdjsr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.3.0"
djoser = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
